P0509R1 Updating "Restrictions on exception handling"
This one is a stinker; it appears to be based on an earlier revision of
p0003.

The new wording in paragraph 1 makes paragraph 2 ("A function may throw
an object of a type not listed in its Throws clause [...]") redundant,
and that paragraph is not listed in the paper (as if it were already
gone), so I have removed it.

The removed wording in paragraph 1 was already removed by p0003r5, so
that change was not applied.

The first sentence of deleted text in paragraph 4 does not exactly match
the text either before or after this paragraph was modified by p0003r5,
but the intent is clear. The second sentence was already deleted by
p0003r5, so that change was not applied.

The added paragraph 5 was already added by p0003r5, so that change was
not applied. (Note that p0003r5 uses "for a" where this paper uses "of
a", but the meaning is the same either way.)

Fixes NB GB 41, GB 42 (C++17 CD)